Provide ser/deser methods
Do not expose ObjectMapper to public, but provide ser/deser API instead to avoid leak of abstraction.
