+++ presto-history-extension/src/test/resources/com/facebook/presto/server/extension/query/history/query-info.json
